Ensure that AndroidInject can match dependencies mocked by Mockito
One of the use cases of AndroidInject is to be able to mock dependencies in tests, and
Mockito is a popular mock testing framework for this.  We thus add a test to ensure
that this case works.

So far mocking an interface and registering it as a dependency with Android-Inject works,
as the framework can match it.  What doesn't work is when an actual class is mocked, as
CGLIB must be used to mock it and thus Android-Inject could no longer match.  Will still
need to figure out how to deal with that scenario.

Signed-off-by: Ealden Esto E. Escanan <ealden@teamcodeflux.com>